+++ plugins/base/src/main/java/sysmon/plugins/base/BaseNetworkExtension.java
@@ -23,7 +23,7 @@ public class BaseNetworkExtension implements MetricExtension {
     // Configuration / Options
     private boolean enabled = true;
     private boolean threaded = false;
-    private String interval = "10s";
+    private String interval = "30s";
 
     private HardwareAbstractionLayer hardwareAbstractionLayer;
     private List<NetworkIF> interfaces;